+++ tests/Feature/Console/IdeMeta/Report/ReportFacade.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Feature\Console\IdeMeta\Report;
+
+use Gacela\Framework\AbstractFacade;
+
+final class ReportFacade extends AbstractFacade
+{
+}