+++ tests/Unit/Framework/ServiceResolver/Fixtures/SecondService.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Unit\Framework\ServiceResolver\Fixtures;
+
+final class SecondService
+{
+}